requirements.txt
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.37.1
django-filter==2.0.0
django-rest-auth==0.9.3
django-rest-swagger==2.0.7
djangorestframework==3.8.2
httpie==0.9.9
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
pbr==4.0.3
Pillow==5.3.0
psycopg2==2.7.5
Pygments==2.2.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.9.0
stevedore==1.28.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==15.2.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
